+++ common/client-libs/validator-client/Cargo.toml
@@ -30,13 +30,12 @@ sha2 = { version = "0.9.5", optional = true }
 itertools = { version = "0.10", optional = true }
 cosmwasm-std = { git = "https://github.com/jstuczyn/cosmwasm", branch="0.14.1-updatedk256", optional = true }
 
-
-[dev-dependencies]
-tokio = {version = "1.5", features = ["full"]}
-
 [target."cfg(target_arch = \"wasm32\")".dependencies.getrandom]
 version = "0.2"
 optional = true
+
+[dev-dependencies]
+tokio = {version = "1.5", features = ["full"]}
 
 [features]
 default = ["nymd-client"]
