+++ common/client-libs/validator-client/src/nymd/mod.rs
@@ -4,20 +4,27 @@
 pub mod cosmwasm_client;
 pub mod wallet;
 
-// struct ValidatorClient {
-//     // ValidatorApiClient and NymdClient
-// }
-//
-// struct ValidatorApiClient {
-//     // whatever, reqwest client probably
-// }
-//
-// struct NymdClient {
-//     // SigningCosmosClient or CosmosClient
-// }
-//
-// struct SigningCosmosClient {}
-//
-// struct CosmosClient {
-//     // http client
-// }
+use crate::nymd::cosmwasm_client::client::CosmWasmClient;
+use crate::nymd::cosmwasm_client::signing_client::SigningCosmWasmClient;
+use crate::nymd::wallet::DirectSecp256k1HdWallet;
+use crate::ValidatorClientError;
+use cosmos_sdk::rpc::{Error as TendermintRpcError, HttpClientUrl};
+use std::convert::TryInto;
+
+pub fn connect<U>(endpoint: U) -> Result<CosmWasmClient, ValidatorClientError>
+where
+    U: TryInto<HttpClientUrl, Error = TendermintRpcError>,
+{
+    CosmWasmClient::connect(endpoint)
+}
+
+// maybe the wallet could be made into a generic, but for now, let's just have this one implementation
+pub fn connect_with_signer<U>(
+    endpoint: U,
+    signer: DirectSecp256k1HdWallet,
+) -> Result<SigningCosmWasmClient, ValidatorClientError>
+where
+    U: TryInto<HttpClientUrl, Error = TendermintRpcError>,
+{
+    SigningCosmWasmClient::connect_with_signer(endpoint, signer)
+}
